Photon keeps near nominal track rather than to direction.
The function nextEdge() has been changed to prefer an edge which keeps the path through the model near the nominal track of the photon rather than an edge whose direction is as near as possible to the photon direction.

Actually 2 versions of the new function have been provided, one which makes use of precalculated values and one which doesn't.

I also included a fix for #189.