Modified getNextEdge() to be more accurate
The algorithm in this function now truly does calculate the distance between the proposed next grid point and the track of the photon (thanks Tuomas Lunttila for suggested code).

I also changed the name of veloproject() to dotProduct3D(), made both its arguments pointers to double, and placed the new function in aux.c.