New parameter collPartMolWeights added to help calculate dust opacity.
This closes #178. LIME until now has used a hard-wired value of 2.4 in its calculation of the dust density which used to take place within calcGridContDustOpacity() and calcGridLinesDustOpacity() (but now in calcDustData(), see below). This number represents the average bulk-gas molecular mass and is the value one would get for a mixture of 80% H2 and 20% He. LIME still defaults to this value if neither par->collPartNames nor par->dustWeights are provided by the user, but the user now has the ability to make use of an average-mass formula appropriate to their particular gas mixture. Note that the default if par->collPartNames are supplied but par->dustWeights not has changed slightly: in this case the exact value for average molecular mass (spatially resolved) is obtained via a simple sum of terms, the ith term being the product of the ith density function value times the molecular mass of the associated species. Previously the densities were simply summed and multiplied by the value of 2.4.

The interplay between the (now 5) collision-partner-related list parameters is delicate, so I have tried to make things clearer by some code rearrangements.

  - New module collparts.c which contains checkUserDensWeights() (moved from aux.c) and setUpDensityAux() (moved from molinit.c).

  - calcGridContDustOpacity() and calcGridLinesDustOpacity() now call the same function calcDustData() to calculate the dust opacity.

Other changes:
  - Moved planckfunc() from molinit.c to aux.c

  - Remaining functions in molinit.c rearranged in order so as to allow removal of their templates from lime.h

  - The CP macros moved to a new include file collparts.h.

  - Some more initialization is now done in mallocAndSetDefaultGrid().

  - Rearranged the elements of structs 'inputPars' and 'configInfo' to make them a bit easier to read. This may have a slight impact on the memory footprint of these structs - this is something which probably ought to be looked at some time.