Slight speedup of raytracing.
In tracerasy_smooth() the grid velocity is interpolated within the cell. This makes use of some buffer space. Previously this was being allocated at each call of the function and some unchanging indices were also calculated with each call. The buffer allocation and the unchanging calculation have now been moved outside the loop over rays.

Note that this also fixes the compiler warnings mentioned in #222.